+++ packages/moonstone/MoonstoneDecorator/MoonstoneDecorator.js
@@ -44,7 +44,6 @@ const defaultConfig = {
  * Moonstone theming to an application. It also applies
  * [floating layer]{@link ui/FloatingLayer.FloatingLayerDecorator},
  * [resolution independence]{@link ui/resolution.ResolutionDecorator},
- * [custom text sizing]{@link moonstone/MoonstoneDecorator.TextSizeDecorator},
  * [skin support]{@link ui/Skinnable}, [spotlight]{@link spotlight.SpotlightRootDecorator}, and
  * [internationalization support]{@link i18n/I18nDecorator.I18nDecorator}. It is meant to be applied to
  * the root element of an app.