+++ packages/spotlight/SpotlightRootDecorator/SpotlightRootDecorator.js
@@ -96,6 +96,5 @@ const SpotlightRootDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 export default SpotlightRootDecorator;
 export {
-	rootContainerId as spotlightRootContainerName, // DEPRECATED
 	SpotlightRootDecorator
 };