+++ packages/i18n/Uppercase/Uppercase.js
@@ -4,13 +4,25 @@
  * @module i18n/Uppercase
  */
 
-import deprecate from '@enact/core/internal/deprecate';
 import hoc from '@enact/core/hoc';
 import kind from '@enact/core/kind';
 import React from 'react';
 import PropTypes from 'prop-types';
 
 import {toCapitalized, toUpperCase, toWordCase} from '../util';
+
+const isString = (content) => typeof content === 'string';
+
+const formatContent = (casing, content) => {
+	switch (casing) {
+		case 'word':
+			return toWordCase(content);
+		case 'sentence':
+			return toCapitalized(content);
+		case 'upper':
+			return toUpperCase(content);
+	}
+};
 
 /**
  * {@link i18n/Uppercase.Uppercase} is a Higher Order Component that is used to wrap
@@ -40,51 +52,26 @@ const Uppercase = hoc((config, Wrapped) => kind({
 		 * @default 'upper'
 		 * @public
 		 */
-		casing: PropTypes.oneOf(['upper', 'preserve', 'word', 'sentence']),
-
-		/**
-		 * The children string will be uppercased, unless this is set to true.
-		 *
-		 * @type {Boolean}
-		 * @default false
-		 * @deprecated replaced by `casing`
-		 * @public
-		 */
-		preserveCase: PropTypes.bool
+		casing: PropTypes.oneOf(['upper', 'preserve', 'word', 'sentence'])
 	},
 
 	defaultProps: {
-		casing: 'upper',
-		preserveCase: false
+		casing: 'upper'
 	},
 
 	computed: {
-		children: ({casing, children, preserveCase}) => {
-			if (preserveCase) {
-				deprecate({name: 'preserveCase', since: '1.1.0', replacedBy: 'casing'});
-				casing = casing || preserveCase && 'preserve';
+		children: ({casing, children}) => {
+			if (casing !== 'preserve' && React.Children.count(children)) {
+				return isString(children) ? formatContent(casing, children) : React.Children.map(children, (child) => {
+					return isString(child) ? formatContent(casing, child) : child;
+				});
 			}
-
-			return (casing !== 'preserve') ? React.Children.map(children, (child) => {
-				if (typeof child == 'string') {
-					switch (casing) {
-						case 'word':
-							return toWordCase(child);
-						case 'sentence':
-							return toCapitalized(child);
-						case 'upper':
-							return toUpperCase(child);
-					}
-				} else {
-					return child;
-				}
-			}) : children;
+			return children;
 		}
 	},
 
 	render: (props) => {
 		delete props.casing;
-		delete props.preserveCase;
 		return (
 			<Wrapped {...props} />
 		);