+++ packages/i18n/CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact i18n module, newest chan
 
 ### Removed
 
-- `i18n/UpperCase` property `preserveCase`, replaced by `casing`
+- `i18n/Uppercase` property `preserveCase`, replaced by `casing`
 
 ## [2.0.0-beta.6] - 2018-06-04